requirements.txt
asgiref==3.3.4
Django==3.2.3
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==4.5.0
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.4.1
